refactor agents to async
