+++ agents/agents/ba_agent.py
@@ -1,20 +1,9 @@
-from settings import llm, graph_maker_prompt
+from settings import llm, ba_prompt, ba_instruction
 from utils.cutomLogger import customLogger
-
-from langchain.agents import initialize_agent, AgentType, Tool
-from langchain.memory import ConversationBufferMemory
-from langgraph.graph import StateGraph
-from langchain.chat_models import ChatOpenAI
-from langchain_gigachat.chat_models import GigaChat
 from langchain.schema import HumanMessage, SystemMessage, Document, AIMessage
-from langchain.vectorstores import FAISS
-from langchain.embeddings import SentenceTransformerEmbeddings
 from langchain.chains import RetrievalQA
-from langchain_community.tools import DuckDuckGoSearchRun
-from langgraph.types import Command
 from langgraph.prebuilt import create_react_agent
 from langgraph.checkpoint.memory import MemorySaver
-from langgraph.graph import StateGraph, END, MessagesState, START
 from typing import TypedDict, Optional, Literal, List, Dict
 from typing_extensions import TypedDict
 
@@ -24,43 +13,45 @@ import os
 import re
 
 _LOGGER = customLogger.getLogger(__name__)
-POSTFIX = "\n\nИсправь, пожалуйста, и сгенерируй граф связей заново."
+POSTFIX = "\n\nИсправь, пожалуйста, и сгенерируй бизнес-требования заново."
 
-class GraphAgentState(TypedDict):
+
+class BaAgentState(TypedDict):
     task: str
-    description: str
     messages: List
     result: str
     questions: List
 
-class GraphAgent:
+class BaAgent:
     def __init__(self):
-        self.graph_agent = self.create_qa_agent()
+        self.agent = self.create_qa_agent()
 
     def create_qa_agent(self):
-        ba_agent = create_react_agent(llm, tools=[], checkpointer=MemorySaver())
-        return ba_agent
+        agent = create_react_agent(llm, tools=[], checkpointer=MemorySaver())
+        return agent
 
-    def run_qa_agent(self, state:GraphAgentState, config:dict):
-        _LOGGER.info(f"Status: graph_agent_node, thread_id: {config['configurable']['thread_id']}")
+    async def run_qa_agent(self, state:BaAgentState, config:dict):
+        _LOGGER.info(f"Status: ba_agent_node, thread_id: {config['configurable']['thread_id']}")
         state["questions"] = ""
+
         if "messages" in state and state["messages"]:
             old_messages = state["messages"]
             request = state["messages"][-1].content + POSTFIX
         else:
             old_messages = []
             request = state["task"]
-            _LOGGER.info(f"GRAPH REQUEST: {request}")
+
+        _LOGGER.info(f"BA REQUEST: {request}")
 
         request = {
             "messages": [HumanMessage(content=request)]
         }
-        response = self.graph_agent.invoke(request, config=config)
+        response = await self.agent.ainvoke(request, config=config)
         if isinstance(response, dict):
             result = response["messages"][-1].content
         else:
             result = response
-        _LOGGER.info(f"GRAPH RESPONSE: {result}")
+        _LOGGER.info(f"BA RESPONSE: {result}")
         matches = re.findall(r'\[ВОПРОС\](.*?)\[/ВОПРОС\]', result, re.DOTALL)
         if matches:
             ques_string = [
@@ -73,7 +64,7 @@ class GraphAgent:
         state["messages"] =  old_messages + [HumanMessage(content=result, name="Аналитик")]
         return state
 
-    def add_message(self, state:GraphAgentState, msg:str):
+    def add_message(self, state:BaAgentState, msg:str):
         if "messages" in state and state["messages"]:
             old_messages = state["messages"]
         else:
@@ -81,12 +72,11 @@ class GraphAgent:
         state["messages"] = old_messages + [HumanMessage(content=msg)]
         return state
 
-    def get_result(self, state:GraphAgentState):
+    def get_result(self, state:BaAgentState):
         if "result" in state and state["result"]:
             return {
                 "status": "OK",
-                "content": state["result"],
-                "mermaid": self.get_mermaid(state["result"])
+                "content": state["result"]
             }
         elif "questions" in state and state["questions"]:
             return {
@@ -99,23 +89,18 @@ class GraphAgent:
                 "content": "Возникла ошибка"
             }
 
-    def run(self, msg: str, state:GraphAgentState, config:dict):
+    async def run(self, msg: str, state:BaAgentState, config:dict):
         if not "messages" in state or not state["messages"]:
-            state["task"] = graph_maker_prompt.format(task=state["task"], description=state["description"])
+            state["task"] = ba_prompt.format(
+                task=state["task"],
+                ba_instruction=ba_instruction
+            )
         else:
             state = self.add_message(state, msg)
-        state = self.run_qa_agent(state, config)
+        state = await self.run_qa_agent(state, config)
         response = self.get_result(state)
         response["state"] = state
         return response
-
-    def get_mermaid(self, result):
-        matches = re.search(r'```mermaid(.*?)```', result, re.DOTALL)
-        if matches:
-            return matches.group(1)
-        else:
-            return None
-
