+++ agents/agents/graph_agent.py
@@ -1,9 +1,7 @@
-from settings import llm, describer_prompt, instruments
+from settings import llm, graph_maker_prompt
 from utils.cutomLogger import customLogger
 from langchain.schema import HumanMessage, SystemMessage, Document, AIMessage
-
 from langchain.chains import RetrievalQA
-
 from langgraph.prebuilt import create_react_agent
 from langgraph.checkpoint.memory import MemorySaver
 from typing import TypedDict, Optional, Literal, List, Dict
@@ -15,43 +13,43 @@ import os
 import re
 
 _LOGGER = customLogger.getLogger(__name__)
-POSTFIX = "\n\nИсправь, пожалуйста, и сгенерируй описание заново."
+POSTFIX = "\n\nИсправь, пожалуйста, и сгенерируй граф связей заново."
 
-
-class DescAgentState(TypedDict):
+class GraphAgentState(TypedDict):
     task: str
+    description: str
     messages: List
-    questions: str
     result: str
+    questions: List
 
-class DescAgent:
+class GraphAgent:
     def __init__(self):
-        self.desc_agent = self.create_qa_agent()
+        self.agent = self.create_qa_agent()
 
     def create_qa_agent(self):
-        desc_agent = create_react_agent(llm, tools=[], checkpointer=MemorySaver())
-        return desc_agent
+        agent = create_react_agent(llm, tools=[], checkpointer=MemorySaver())
+        return agent
 
-    def run_qa_agent(self, state:DescAgentState, config:dict):
-        _LOGGER.info(f"Status: desc_agent_node, thread_id: {config['configurable']['thread_id']}")
+    async def run_qa_agent(self, state:GraphAgentState, config:dict):
+        _LOGGER.info(f"Status: graph_agent_node, thread_id: {config['configurable']['thread_id']}")
         state["questions"] = ""
         if "messages" in state and state["messages"]:
             old_messages = state["messages"]
             request = state["messages"][-1].content + POSTFIX
-            _LOGGER.info(f"REQUEST: {request}")
-
         else:
             old_messages = []
             request = state["task"]
+        _LOGGER.info(f"GRAPH REQUEST: {request}")
+
         request = {
             "messages": [HumanMessage(content=request)]
         }
-        response = self.desc_agent.invoke(request, config=config)
+        response = await self.agent.ainvoke(request, config=config)
         if isinstance(response, dict):
             result = response["messages"][-1].content
         else:
             result = response
-        _LOGGER.info(f"DESC RESPONSE: {result}")
+        _LOGGER.info(f"GRAPH RESPONSE: {result}")
         matches = re.findall(r'\[ВОПРОС\](.*?)\[/ВОПРОС\]', result, re.DOTALL)
         if matches:
             ques_string = [
@@ -64,7 +62,7 @@ class DescAgent:
         state["messages"] =  old_messages + [HumanMessage(content=result, name="Аналитик")]
         return state
 
-    def add_message(self, state:DescAgentState, msg:str):
+    def add_message(self, state:GraphAgentState, msg:str):
         if "messages" in state and state["messages"]:
             old_messages = state["messages"]
         else:
@@ -72,11 +70,12 @@ class DescAgent:
         state["messages"] = old_messages + [HumanMessage(content=msg)]
         return state
 
-    def get_result(self, state:DescAgentState):
+    def get_result(self, state:GraphAgentState):
         if "result" in state and state["result"]:
             return {
                 "status": "OK",
-                "content": state["result"]
+                "content": state["result"],
+                "mermaid": self.get_mermaid(state["result"])
             }
         elif "questions" in state and state["questions"]:
             return {
@@ -89,15 +88,23 @@ class DescAgent:
                 "content": "Возникла ошибка"
             }
 
-    def run(self, msg: str, state:DescAgentState, config:dict):
+    async def run(self, msg: str, state:GraphAgentState, config:dict):
         if not "messages" in state or not state["messages"]:
-            state["task"] = describer_prompt.format(task=msg, instruments=instruments)
+            state["task"] = graph_maker_prompt.format(task=state["task"], description=state["description"])
         else:
             state = self.add_message(state, msg)
-        state = self.run_qa_agent(state, config)
+        state = await self.run_qa_agent(state, config)
         response = self.get_result(state)
         response["state"] = state
         return response
+
+    def get_mermaid(self, result):
+        matches = re.search(r'```mermaid(.*?)```', result, re.DOTALL)
+        if matches:
+            return matches.group(1)
+        else:
+            return None
+
